# slightly decrease thread contention by synchronizing once rather than…