Improve Summary: slightly decrease thread contention by synchronizing once rather than "number quantiles + 1" times