@rrakos-evo 2019-06-03
It looks like this doesn't make a significant difference. 🙁 

Benchmark:
```
# JMH version: 1.21
# VM version: JDK 11.0.2, OpenJDK 64-Bit Server VM, 11.0.2+9
# VM invoker: /Library/Java/JavaVirtualMachines/openjdk-11.0.2.jdk/Contents/Home/bin/java
# VM options: -javaagent:/Applications/IntelliJ IDEA CE.app/Contents/lib/idea_rt.jar=49590:/Applications/IntelliJ IDEA CE.app/Contents/bin -Dfile.encoding=UTF-8
# Warmup: 5 iterations, 10 s each
# Measurement: 4 iterations, 10 s each
# Timeout: 10 min per iteration
# Threads: 24 threads, will synchronize iterations
```
Please take the results with a grain of salt.

Before - 4e0e7527:
```
Benchmark                                                           Mode  Cnt       Score        Error  Units
SummaryBenchmark.codahaleHistogramBenchmark                         avgt    4    5449.211 ±    157.085  ns/op
SummaryBenchmark.prometheusSimpleHistogramBenchmark                 avgt    4     219.256 ±     23.811  ns/op
SummaryBenchmark.prometheusSimpleHistogramChildBenchmark            avgt    4      88.585 ±      1.734  ns/op
SummaryBenchmark.prometheusSimpleHistogramNoLabelsBenchmark         avgt    4      83.989 ±      2.093  ns/op
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     186.796 ±     15.029  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      61.782 ±      5.032  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      59.388 ±     12.457  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  530692.958 ± 627614.926  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  563807.323 ± 389413.917  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  559719.005 ± 631676.260  ns/op
```

After - c726b1a7:
```
Benchmark                                                           Mode  Cnt       Score        Error  Units
SummaryBenchmark.codahaleHistogramBenchmark                         avgt    4    5651.711 ±    487.476  ns/op
SummaryBenchmark.prometheusSimpleHistogramBenchmark                 avgt    4     220.865 ±     11.554  ns/op
SummaryBenchmark.prometheusSimpleHistogramChildBenchmark            avgt    4      95.237 ±      6.059  ns/op
SummaryBenchmark.prometheusSimpleHistogramNoLabelsBenchmark         avgt    4      87.088 ±     15.551  ns/op
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     196.351 ±     33.766  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      60.853 ±      9.388  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      64.209 ±      2.437  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  525567.096 ± 599616.355  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  724240.526 ± 864857.588  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  703944.682 ± 906458.910  ns/op
```
I picked the result with the smallest error from multiple benchmark runs.

Quantiles:
```
  .quantile(0.5, 0.05).quantile(0.9, 0.05).quantile(0.95, 0.01).quantile(0.99, 0.005)
```