@brian-brazil 2019-06-03
I still have to do a proper review, however the benchmarks are not
conclusive. Try a benchmark that rotates frequently.

On Mon 3 Jun 2019, 17:36 Yaroslav Klymko, <notifications@github.com> wrote:

> @brian-brazil <https://github.com/brian-brazil> is there something which
> stops you from accepting this PR ? how can I improve this?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/prometheus/client_java/pull/481?email_source=notifications&email_token=ABWJG5SCLH5D2V7NVPSGZL3PYU3AJA5CNFSM4HSDEZSKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODWZZPCI#issuecomment-498309001>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABWJG5RIOIPORX63JOE6JQTPYU3AJANCNFSM4HSDEZSA>
> .
>