@rrakos-evo 2019-06-04
It looks like this doesn't make a significant difference. 🙁

Before - 4e0e752 - `maxAgeSeconds=600, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score         Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     219.757 ±      61.084  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      81.192 ±      42.157  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      71.233 ±      10.217  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  544183.558 ±  673721.964  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  604961.225 ± 1102601.976  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  611718.456 ±  961269.567  ns/op
```

After - c726b1a - `maxAgeSeconds=600, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score         Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     252.022 ±     161.467  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      69.318 ±       9.407  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      70.984 ±       8.715  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  508342.249 ±  367462.250  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  661664.689 ±  739166.750  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  697726.672 ± 1409700.286  ns/op
```

Before - 4e0e752 - `maxAgeSeconds=60, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score        Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     213.747 ±      7.583  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      66.990 ±      6.503  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      77.389 ±      9.192  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  259765.997 ± 358709.484  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  317509.004 ± 596762.748  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  324706.634 ± 382327.655  ns/op
```

After - c726b1a - `maxAgeSeconds=60, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score        Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     231.908 ±     45.979  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      74.272 ±      8.909  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      72.157 ±      4.310  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  352584.198 ± 191647.019  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  411203.918 ± 596264.924  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  410841.897 ± 505489.988  ns/op
```

Before - 4e0e752 - `maxAgeSeconds=6, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score       Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     246.210 ±    17.015  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      67.433 ±    10.148  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      70.137 ±    10.208  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  106283.964 ± 59817.072  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  117276.079 ± 15151.896  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  112516.505 ± 48298.023  ns/op
```

After - c726b1a - `maxAgeSeconds=6, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt       Score       Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4     225.264 ±    57.002  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4      64.463 ±     2.153  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4      66.457 ±     1.970  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  114767.996 ± 40131.263  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  126502.374 ± 30720.763  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  131050.590 ± 65970.924  ns/op
```

Before - 4e0e752 - `maxAgeSeconds=1, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt      Score       Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4    226.792 ±    51.313  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4     67.871 ±    10.559  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4     68.260 ±     6.567  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  45594.702 ± 12321.879  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  48994.075 ±  3317.460  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  47073.086 ±  5388.792  ns/op
```

After - c726b1a - `maxAgeSeconds=1, ageBuckets=5`:
```
Benchmark                                                           Mode  Cnt      Score       Error  Units
SummaryBenchmark.prometheusSimpleSummaryBenchmark                   avgt    4    213.242 ±    51.151  ns/op
SummaryBenchmark.prometheusSimpleSummaryChildBenchmark              avgt    4     67.623 ±    20.423  ns/op
SummaryBenchmark.prometheusSimpleSummaryNoLabelsBenchmark           avgt    4     69.312 ±     8.655  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesBenchmark          avgt    4  46977.894 ± 21980.921  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesChildBenchmark     avgt    4  44708.777 ± 11136.298  ns/op
SummaryBenchmark.prometheusSimpleSummaryQuantilesNoLabelsBenchmark  avgt    4  45772.478 ±  7892.974  ns/op
```

It makes sense to me that more frequent rotations (shorter time window) is faster, as there are less measurements and samples (for `CKMSQuantiles`).